Planner: move properties in shared memory

+++ cogip/cpp/libraries/utils/include/utils/LidarDataConverter.hpp
@@ -37,17 +37,17 @@ public:
     }
 
     /// Set the table limits margin.
-    void setTableLimitsMargin(float table_limits_margin) {
+    void setTableLimitsMargin(double table_limits_margin) {
         table_limits_margin_ = table_limits_margin;
     }
 
     /// Set the lidar offset on X axis.
-    void setLidarOffsetX(float lidar_offset_x) {
+    void setLidarOffsetX(double lidar_offset_x) {
         lidar_offset_x_ = lidar_offset_x;
     }
 
     /// Set the lidar offset on Y axis.
-    void setLidarOffsetY(float lidar_offset_y) {
+    void setLidarOffsetY(double lidar_offset_y) {
         lidar_offset_y_ = lidar_offset_y;
     }
 
@@ -58,16 +58,16 @@ public:
 
 private:
     shared_memory::SharedMemory shared_memory_;                   ///< Shared memory instance
-    float (*lidar_data_)[3];                                      ///< Pointer to lidar data memory
-    float (*lidar_coords_)[2];                                    ///< Pointer to lidar coords memory
+    double (*lidar_data_)[3];                                     ///< Pointer to lidar data memory
+    double (*lidar_coords_)[2];                                   ///< Pointer to lidar coords memory
     cogip::shared_memory::WritePriorityLock data_read_lock_;      ///< Lock for reading lidar data
     cogip::shared_memory::WritePriorityLock coords_write_lock_;   ///< Lock for writing lidar coordinates
     models::PoseBuffer* pose_current_buffer_;                     ///< Pointer to the current pose buffer
     std::size_t pose_current_index_;                              ///< Index of the current pose
-    float* table_limits_;                                         ///< Pointer to table limits
-    float table_limits_margin_;                                   ///< Margin for table limits
-    float lidar_offset_x_;                                        ///< Lidar offset on X axis
-    float lidar_offset_y_;                                        ///< Lidar offset on Y axis
+    double* table_limits_;                                        ///< Pointer to table limits
+    double table_limits_margin_;                                  ///< Margin for table limits
+    double lidar_offset_x_;                                       ///< Lidar offset on X axis
+    double lidar_offset_y_;                                       ///< Lidar offset on Y axis
     bool running_;                                                ///< Flag to indicate if the converter is running
     std::thread thread_;                                          ///< Thread for the converter
     bool debug_;                                                  ///< Flag to enable debug mode
